models/canny2seg/pix2pix/requirements.txt
torch>=2.1.0
torchvision>=0.16.0
numpy>=1.24.0
pillow>=10.0.0
tqdm>=4.66.0
pyyaml>=6.0.1
kaggle>=1.6.14
